src/main/java/guru/springframework/spring5receipeapp/services/RecipeServiceImpl.java
package guru.springframework.spring5receipeapp.services;

import guru.springframework.spring5receipeapp.domain.Recipe;
import guru.springframework.spring5receipeapp.repositiories.RecipeRepository;
import org.springframework.stereotype.Service;

import java.util.HashSet;
import java.util.Set;

@Service
public class RecipeServiceImpl implements RecipeService {

    private final RecipeRepository recipeRepository;

    public RecipeServiceImpl(RecipeRepository recipeRepository) {
        this.recipeRepository = recipeRepository;
    }

    @Override
    public Set<Recipe> getReceipes() {
        Set<Recipe> recipes=new HashSet<>();
        recipeRepository.findAll().iterator().forEachRemaining(recipes::add);
        return recipes;
    }
}
